fix(clippy): reduce warnings from 73 to 14
- Fix unused seq_len parameter (prefix with _)
- Fix needless_range_loop in preflight.rs (use iterator)
- Fix explicit_counter_loop in inference.rs (use enumerate)
- Add allow for legitimate index-based loop in frontier_compare
- Fix libfse manual_div_ceil (use .div_ceil())
- Add type aliases: InferenceResult, LayerDebugOutput

+++ tests/template_pipeline.rs
@@ -92,7 +92,7 @@ fn read_tokenizer_info(path: &Path) -> Option<GgufTokenizerInfo> {
     let n_kv = if version >= 2 {
         read_u64(&mut f).ok()?; read_u64(&mut f).ok()?
     } else {
-        read_u32(&mut f).ok()? as u64; read_u32(&mut f).ok()? as u64
+        read_u32(&mut f).ok()?; read_u32(&mut f).ok()? as u64
     };
 
     let mut info = GgufTokenizerInfo {
